+++ packages/camera/camera_avfoundation/ios/camera_avfoundation/Sources/camera_avfoundation/DefaultCamera.swift
@@ -10,6 +10,14 @@ import CoreMotion
 #endif
 
 final class DefaultCamera: FLTCam, Camera {
+  override var videoFormat: FourCharCode {
+    didSet {
+      captureVideoOutput.videoSettings = [
+        kCVPixelBufferPixelFormatTypeKey as String: videoFormat
+      ]
+    }
+  }
+
   override var deviceOrientation: UIDeviceOrientation {
     get { super.deviceOrientation }
     set {
@@ -50,6 +58,34 @@ final class DefaultCamera: FLTCam, Camera {
       code: "Error \(error.code)",
       message: error.localizedDescription,
       details: error.domain)
+  }
+
+  private static func createConnection(
+    captureDevice: FLTCaptureDevice,
+    videoFormat: FourCharCode,
+    captureDeviceInputFactory: FLTCaptureDeviceInputFactory
+  ) throws -> (FLTCaptureInput, FLTCaptureVideoDataOutput, AVCaptureConnection) {
+    // Setup video capture input.
+    let captureVideoInput = try captureDeviceInputFactory.deviceInput(with: captureDevice)
+
+    // Setup video capture output.
+    let captureVideoOutput = FLTDefaultCaptureVideoDataOutput(
+      captureVideoOutput: AVCaptureVideoDataOutput())
+    captureVideoOutput.videoSettings = [
+      kCVPixelBufferPixelFormatTypeKey as String: videoFormat
+    ]
+    captureVideoOutput.alwaysDiscardsLateVideoFrames = true
+
+    // Setup video capture connection.
+    let connection = AVCaptureConnection(
+      inputPorts: captureVideoInput.ports,
+      output: captureVideoOutput.avOutput)
+
+    if captureDevice.position == .front {
+      connection.isVideoMirrored = true
+    }
+
+    return (captureVideoInput, captureVideoOutput, connection)
   }
 
   func reportInitializationState() {
@@ -94,6 +130,40 @@ final class DefaultCamera: FLTCam, Camera {
 
   func resumeVideoRecording() {
     isRecordingPaused = false
+  }
+
+  func stopVideoRecording(completion: @escaping (String?, FlutterError?) -> Void) {
+    guard isRecording else {
+      let error = NSError(
+        domain: NSCocoaErrorDomain,
+        code: URLError.resourceUnavailable.rawValue,
+        userInfo: [NSLocalizedDescriptionKey: "Video is not recording!"]
+      )
+      completion(nil, DefaultCamera.flutterErrorFromNSError(error))
+      return
+    }
+
+    isRecording = false
+
+    // When `isRecording` is true `startWriting` was already called so `videoWriter.status`
+    // is always either `.writing` or `.failed` and `finishWriting` does not throw exceptions so
+    // there is no need to check `videoWriter.status`
+    videoWriter?.finishWriting { [weak self] in
+      guard let strongSelf = self else { return }
+
+      if strongSelf.videoWriter?.status == .completed {
+        strongSelf.updateOrientation()
+        completion(strongSelf.videoRecordingPath, nil)
+        strongSelf.videoRecordingPath = nil
+      } else {
+        completion(
+          nil,
+          FlutterError(
+            code: "IOError",
+            message: "AVAssetWriter could not finish writing!",
+            details: nil))
+      }
+    }
   }
 
   func lockCaptureOrientation(_ pigeonOrientation: FCPPlatformDeviceOrientation) {
@@ -339,6 +409,94 @@ final class DefaultCamera: FLTCam, Camera {
 
   func resumePreview() {
     isPreviewPaused = false
+  }
+
+  func setDescriptionWhileRecording(
+    _ cameraName: String, withCompletion completion: @escaping (FlutterError?) -> Void
+  ) {
+    guard isRecording else {
+      completion(
+        FlutterError(
+          code: "setDescriptionWhileRecordingFailed",
+          message: "Device was not recording",
+          details: nil))
+      return
+    }
+
+    captureDevice = captureDeviceFactory(cameraName)
+
+    let oldConnection = captureVideoOutput.connection(withMediaType: .video)
+
+    // Stop video capture from the old output.
+    captureVideoOutput.setSampleBufferDelegate(nil, queue: nil)
+
+    // Remove the old video capture connections.
+    videoCaptureSession.beginConfiguration()
+    videoCaptureSession.removeInput(captureVideoInput)
+    videoCaptureSession.removeOutput(captureVideoOutput.avOutput)
+
+    let newConnection: AVCaptureConnection
+
+    do {
+      (captureVideoInput, captureVideoOutput, newConnection) = try DefaultCamera.createConnection(
+        captureDevice: captureDevice,
+        videoFormat: videoFormat,
+        captureDeviceInputFactory: captureDeviceInputFactory)
+
+      captureVideoOutput.setSampleBufferDelegate(self, queue: captureSessionQueue)
+    } catch {
+      completion(
+        FlutterError(
+          code: "VideoError",
+          message: "Unable to create video connection",
+          details: nil))
+      return
+    }
+
+    // Keep the same orientation the old connections had.
+    if let oldConnection = oldConnection, newConnection.isVideoOrientationSupported {
+      newConnection.videoOrientation = oldConnection.videoOrientation
+    }
+
+    // Add the new connections to the session.
+    if !videoCaptureSession.canAddInput(captureVideoInput) {
+      completion(
+        FlutterError(
+          code: "VideoError",
+          message: "Unable to switch video input",
+          details: nil))
+    }
+    videoCaptureSession.addInputWithNoConnections(captureVideoInput)
+
+    if !videoCaptureSession.canAddOutput(captureVideoOutput.avOutput) {
+      completion(
+        FlutterError(
+          code: "VideoError",
+          message: "Unable to switch video output",
+          details: nil))
+    }
+    videoCaptureSession.addOutputWithNoConnections(captureVideoOutput.avOutput)
+
+    if !videoCaptureSession.canAddConnection(newConnection) {
+      completion(
+        FlutterError(
+          code: "VideoError",
+          message: "Unable to switch video connection",
+          details: nil))
+    }
+    videoCaptureSession.addConnection(newConnection)
+    videoCaptureSession.commitConfiguration()
+
+    completion(nil)
+  }
+
+  func stopImageStream() {
+    if isStreamingImages {
+      isStreamingImages = false
+      imageStreamHandler = nil
+    } else {
+      reportErrorMessage("Images from camera are not streaming!")
+    }
   }
 
   func captureOutput(
@@ -590,5 +748,9 @@ final class DefaultCamera: FLTCam, Camera {
         // Ignore any errors, as this is just an event broadcast.
       }
     }
+  }
+
+  deinit {
+    motionManager.stopAccelerometerUpdates()
   }
 }