+++ packages/camera/camera_avfoundation/ios/camera_avfoundation/Sources/camera_avfoundation_objc/FLTCam.m
@@ -5,7 +5,6 @@
 #import "./include/camera_avfoundation/FLTCam.h"
 #import "./include/camera_avfoundation/FLTCam_Test.h"
 
-@import CoreMotion;
 @import Flutter;
 #import <libkern/OSAtomic.h>
 
@@ -33,18 +32,13 @@ static FlutterError *FlutterErrorFromNSError(NSError *error) {
 @property(readonly, nonatomic) FCPPlatformMediaSettings *mediaSettings;
 @property(readonly, nonatomic) FLTCamMediaSettingsAVWrapper *mediaSettingsAVWrapper;
 
-@property(readonly, nonatomic) NSObject<FLTCaptureInput> *captureVideoInput;
 @property(readonly, nonatomic) CGSize captureSize;
 @property(strong, nonatomic)
     NSObject<FLTAssetWriterInputPixelBufferAdaptor> *assetWriterPixelBufferAdaptor;
 @property(strong, nonatomic) AVCaptureVideoDataOutput *videoOutput;
 @property(strong, nonatomic) AVCaptureAudioDataOutput *audioOutput;
-@property(strong, nonatomic) NSString *videoRecordingPath;
 @property(assign, nonatomic) BOOL isAudioSetup;
 
-@property(nonatomic) CMMotionManager *motionManager;
-/// All FLTCam's state access and capture session related operations should be on run on this queue.
-@property(strong, nonatomic) dispatch_queue_t captureSessionQueue;
 /// The queue on which captured photos (not videos) are written to disk.
 /// Videos are written to disk by `videoAdaptor` on an internal queue managed by AVFoundation.
 @property(strong, nonatomic) dispatch_queue_t photoIOQueue;
@@ -52,9 +46,7 @@ static FlutterError *FlutterErrorFromNSError(NSError *error) {
 /// Allows for alternate implementations in tests.
 @property(nonatomic, copy) VideoDimensionsForFormat videoDimensionsForFormat;
 /// A wrapper for AVCaptureDevice creation to allow for dependency injection in tests.
-@property(nonatomic, copy) CaptureDeviceFactory captureDeviceFactory;
 @property(nonatomic, copy) AudioCaptureDeviceFactory audioCaptureDeviceFactory;
-@property(readonly, nonatomic) NSObject<FLTCaptureDeviceInputFactory> *captureDeviceInputFactory;
 @property(nonatomic, copy) AssetWriterFactory assetWriterFactory;
 @property(nonatomic, copy) InputPixelBufferAdaptorFactory inputPixelBufferAdaptorFactory;
 /// Reports the given error message to the Dart side of the plugin.
@@ -191,12 +183,6 @@ NSString *const errorMethod = @"error";
   }
 
   return connection;
-}
-
-- (void)setVideoFormat:(OSType)videoFormat {
-  _videoFormat = videoFormat;
-  _captureVideoOutput.videoSettings =
-      @{(NSString *)kCVPixelBufferPixelFormatTypeKey : @(videoFormat)};
 }
 
 - (void)updateOrientation {
@@ -426,10 +412,6 @@ NSString *const errorMethod = @"error";
   return bestFormat;
 }
 
-- (void)dealloc {
-  [_motionManager stopAccelerometerUpdates];
-}
-
 /// Main logic to setup the video recording.
 - (void)setUpVideoRecordingWithCompletion:(void (^)(FlutterError *_Nullable))completion {
   NSError *error;
@@ -481,90 +463,6 @@ NSString *const errorMethod = @"error";
   }
 }
 
-- (void)stopVideoRecordingWithCompletion:(void (^)(NSString *_Nullable,
-                                                   FlutterError *_Nullable))completion {
-  if (_isRecording) {
-    _isRecording = NO;
-
-    // when _isRecording is YES startWriting was already called so _videoWriter.status
-    // is always either AVAssetWriterStatusWriting or AVAssetWriterStatusFailed and
-    // finishWritingWithCompletionHandler does not throw exception so there is no need
-    // to check _videoWriter.status
-    [_videoWriter finishWritingWithCompletionHandler:^{
-      if (self->_videoWriter.status == AVAssetWriterStatusCompleted) {
-        [self updateOrientation];
-        completion(self->_videoRecordingPath, nil);
-        self->_videoRecordingPath = nil;
-      } else {
-        completion(nil, [FlutterError errorWithCode:@"IOError"
-                                            message:@"AVAssetWriter could not finish writing!"
-                                            details:nil]);
-      }
-    }];
-  } else {
-    NSError *error =
-        [NSError errorWithDomain:NSCocoaErrorDomain
-                            code:NSURLErrorResourceUnavailable
-                        userInfo:@{NSLocalizedDescriptionKey : @"Video is not recording!"}];
-    completion(nil, FlutterErrorFromNSError(error));
-  }
-}
-
-- (void)setDescriptionWhileRecording:(NSString *)cameraName
-                      withCompletion:(void (^)(FlutterError *_Nullable))completion {
-  if (!_isRecording) {
-    completion([FlutterError errorWithCode:@"setDescriptionWhileRecordingFailed"
-                                   message:@"Device was not recording"
-                                   details:nil]);
-    return;
-  }
-
-  _captureDevice = self.captureDeviceFactory(cameraName);
-
-  NSObject<FLTCaptureConnection> *oldConnection =
-      [_captureVideoOutput connectionWithMediaType:AVMediaTypeVideo];
-
-  // Stop video capture from the old output.
-  [_captureVideoOutput setSampleBufferDelegate:nil queue:nil];
-
-  // Remove the old video capture connections.
-  [_videoCaptureSession beginConfiguration];
-  [_videoCaptureSession removeInput:_captureVideoInput];
-  [_videoCaptureSession removeOutput:_captureVideoOutput.avOutput];
-
-  NSError *error = nil;
-  AVCaptureConnection *newConnection = [self createConnection:&error];
-  if (error) {
-    completion(FlutterErrorFromNSError(error));
-    return;
-  }
-
-  // Keep the same orientation the old connections had.
-  if (oldConnection && newConnection.isVideoOrientationSupported) {
-    newConnection.videoOrientation = oldConnection.videoOrientation;
-  }
-
-  // Add the new connections to the session.
-  if (![_videoCaptureSession canAddInput:_captureVideoInput])
-    completion([FlutterError errorWithCode:@"VideoError"
-                                   message:@"Unable switch video input"
-                                   details:nil]);
-  [_videoCaptureSession addInputWithNoConnections:_captureVideoInput];
-  if (![_videoCaptureSession canAddOutput:_captureVideoOutput.avOutput])
-    completion([FlutterError errorWithCode:@"VideoError"
-                                   message:@"Unable switch video output"
-                                   details:nil]);
-  [_videoCaptureSession addOutputWithNoConnections:_captureVideoOutput.avOutput];
-  if (![_videoCaptureSession canAddConnection:newConnection])
-    completion([FlutterError errorWithCode:@"VideoError"
-                                   message:@"Unable switch video connection"
-                                   details:nil]);
-  [_videoCaptureSession addConnection:newConnection];
-  [_videoCaptureSession commitConfiguration];
-
-  completion(nil);
-}
-
 - (void)startImageStreamWithMessenger:(NSObject<FlutterBinaryMessenger> *)messenger
                            completion:(void (^)(FlutterError *))completion {
   [self startImageStreamWithMessenger:messenger
@@ -609,15 +507,6 @@ NSString *const errorMethod = @"error";
   } else {
     [self reportErrorMessage:@"Images from camera are already streaming!"];
     completion(nil);
-  }
-}
-
-- (void)stopImageStream {
-  if (_isStreamingImages) {
-    _isStreamingImages = NO;
-    _imageStreamHandler = nil;
-  } else {
-    [self reportErrorMessage:@"Images from camera are not streaming!"];
   }
 }